Refactor agent chat panel into modular package
