+++ app/ui/agent_chat_panel/__init__.py
@@ -1,0 +1,13 @@
+"""Agent chat panel package."""
+
+from .execution import AgentCommandExecutor, ThreadedAgentCommandExecutor
+from .panel import AgentChatPanel, RequirementConfirmPreference
+from .paths import history_path_for_documents
+
+__all__ = [
+    "AgentChatPanel",
+    "RequirementConfirmPreference",
+    "AgentCommandExecutor",
+    "ThreadedAgentCommandExecutor",
+    "history_path_for_documents",
+]
